refactor: DB 로직을 OmokRepository로 분리하여 코드 구조 개선
