+++ app/src/main/java/woowacourse/omok/MainActivity.kt
@@ -1,10 +1,7 @@
 package woowacourse.omok
 
-import android.content.ContentValues
 import android.content.Intent
-import android.database.Cursor
 import android.os.Bundle
-import android.util.Log
 import android.widget.ImageView
 import android.widget.TableLayout
 import android.widget.TableRow
@@ -16,8 +13,7 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.children
 import rule.type.Violation
-import woowacourse.omok.domain.db.DbHelper
-import woowacourse.omok.domain.db.OmokContract
+import woowacourse.omok.domain.db.OmokRepository
 import woowacourse.omok.domain.omokboard.ColumnPosition
 import woowacourse.omok.domain.omokboard.OmokBoard
 import woowacourse.omok.domain.omokboard.PlayingBoard
@@ -38,7 +34,7 @@ class MainActivity : AppCompatActivity() {
     private lateinit var board: TableLayout
     private lateinit var playingBoard: PlayingBoard
     private lateinit var omokGame: OmokGame
-    private lateinit var dbHelper: DbHelper
+    private lateinit var omokRepository: OmokRepository
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -53,8 +49,7 @@ class MainActivity : AppCompatActivity() {
         board = findViewById(R.id.board)
         playingBoard = PlayingBoard(OmokBoard(), RuleNavigation(OmokRule.whiteRules, OmokRule.blackRules))
         omokGame = OmokGame(playingBoard)
-
-        dbHelper = DbHelper(this)
+        omokRepository = OmokRepository(this)
 
         board
             .children
@@ -65,7 +60,7 @@ class MainActivity : AppCompatActivity() {
                 val row = index / 15
                 val column = index % 15
 
-                queryStoneAt(row, column)?.let { color ->
+                omokRepository.queryStoneAt(row, column)?.let { color ->
                     val stoneColor = StoneColor.valueOf(color)
                     playingBoard.placeStone(PlayerStone(stoneColor, Position(RowPosition(row + 1), ColumnPosition(column + 1))))
                     stoneColor.toUi(imageView)
@@ -73,66 +68,12 @@ class MainActivity : AppCompatActivity() {
 
                 imageView.setOnClickListener {
                     val currentColor = omokGame.currentStoneColor.name
-                    insertStone(currentColor, row, column)
+                    omokRepository.insertStone(currentColor, row, column)
                     omokGame.start(Position(RowPosition(row + 1), ColumnPosition(column + 1))) { placeResult ->
                         handlePlaceResult(placeResult, imageView)
                     }
                 }
             }
-    }
-
-    private fun insertStone(
-        color: String,
-        row: Int,
-        column: Int,
-    ) {
-        val db = dbHelper.writableDatabase
-
-        val values =
-            ContentValues().apply {
-                put(OmokContract.STONE_COLOR, color)
-                put(OmokContract.POSITION_ROW, row)
-                put(OmokContract.POSITION_COLUMN, column)
-            }
-
-        val newRowId = db.insert(OmokContract.TABLE_NAME, null, values)
-        if (newRowId == -1L) {
-            Log.e("MainActivity", "insert failed")
-        } else {
-            Log.d("MainActivity", "insert success: $newRowId")
-        }
-        db.close()
-    }
-
-    private fun queryStoneAt(
-        row: Int,
-        column: Int,
-    ): String? {
-        val dbReader = dbHelper.readableDatabase
-        var result: String? = null
-
-        val cursor: Cursor =
-            dbReader.query(
-                OmokContract.TABLE_NAME,
-                arrayOf(OmokContract.STONE_COLOR),
-                "${OmokContract.POSITION_ROW} = ? AND ${OmokContract.POSITION_COLUMN} = ?",
-                arrayOf(row.toString(), column.toString()),
-                null,
-                null,
-                null,
-            )
-
-        if (cursor.moveToFirst()) {
-            result = cursor.getString(cursor.getColumnIndexOrThrow(OmokContract.STONE_COLOR))
-        }
-        cursor.close()
-        return result
-    }
-
-    private fun resetDatabase() {
-        val db = dbHelper.writableDatabase
-        db.delete(OmokContract.TABLE_NAME, null, null)
-        db.close()
     }
 
     private fun handlePlaceResult(
@@ -157,7 +98,7 @@ class MainActivity : AppCompatActivity() {
             .setMessage(displayGameResultMessage(gameResult))
             .setPositiveButton("한번 더하기") { dialog, _ ->
                 dialog.dismiss()
-                resetDatabase()
+                omokRepository.resetDatabase()
                 startActivity(Intent(this, MainActivity::class.java))
                 finish()
             }
@@ -212,7 +153,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     override fun onDestroy() {
-        dbHelper.close()
+        omokRepository.close()
 
         super.onDestroy()
     }
